Base Geom ggproto classes for ggplot2

All geom_ functions (like geom_point) return a layer that contains a Geom object (like GeomPoint). The Geom object is responsible for rendering the data in the plot. Each of the Geom objects is a ggproto object, descended from the top-level Geom, and each implements various methods and fields. Compared to Stat and Position, Geom is a little different because the execution of the setup and compute functions is split up. setup_data runs before position adjustments, and draw_layer is not run until render time, much later. This means there is no setup_params because it's hard to communicate the changes.

GeomDensity_

GeomBar_

GeomQuantiles

GeomSerialaxesDensity

GeomSerialaxesHist

GeomSerialaxesQuantile

GeomSerialaxes

Format

An object of class GeomDensity_ (inherits from GeomRibbon, Geom, ggproto, gg) of length 6.

An object of class GeomBar_ (inherits from GeomBar, GeomRect, Geom, ggproto, gg) of length 4.

An object of class GeomQuantiles (inherits from GeomQuantile, GeomPath, Geom, ggproto, gg) of length 1.

An object of class GeomSerialaxesDensity (inherits from GeomDensity_, GeomRibbon, Geom, ggproto, gg) of length 2.

An object of class GeomSerialaxesHist (inherits from GeomBar_, GeomBar, GeomRect, Geom, ggproto, gg) of length 2.

An object of class GeomSerialaxesQuantile (inherits from GeomPath, Geom, ggproto, gg) of length 4.

An object of class GeomSerialaxes (inherits from GeomPath, Geom, ggproto, gg) of length 3.