Base Position ggproto classes for ggplot2

All position_ functions (like position_dodge) return a Position object (like PositionDodge). The Position object is responsible for adjusting the position of overlapping geoms. The way that the position_ functions work is slightly different from the geom_ and stat_ functions, because a position_ function actually "instantiates" the Position object by creating a descendant, and returns that. Each of the Position objects is a ggproto object, descended from the top-level Position.

PositionDodge_

PositionDodge2_

PositionIdentity_

PositionStack_

PositionFill_

Format

An object of class PositionDodge_ (inherits from PositionDodge, Position, ggproto, gg) of length 2.

An object of class PositionDodge2_ (inherits from PositionDodge2, PositionDodge, Position, ggproto, gg) of length 2.

An object of class PositionIdentity_ (inherits from PositionIdentity, Position, ggproto, gg) of length 3.

An object of class PositionStack_ (inherits from PositionStack, Position, ggproto, gg) of length 3.

An object of class PositionFill_ (inherits from PositionStack_, PositionStack, Position, ggproto, gg) of length 2.